Update Slack Github Action condition
The condition for the Slack Github Action in the action.yml file was previously tied to a deployment input check. This commit removes the deployment check, and the Slack Github Action will be invoked on all occasions irrespective of deployment status. This provides more flexibility and ensures notifications are always sent.